Tune strategy 4 further
- Replace multiplicative sim calculation with additive one
- Extract genre BR filtering to helper
